Introduce Transaction new types
This allows us to differentiate between Redeem, Refund and Funding
transactions and avoids errors down the line where they could be mixed
up accidentally.
