+++ application/comit_node/src/swap_protocols/rfc003/ledger.rs
@@ -5,7 +5,7 @@ use swap_protocols::{
     rfc003::secret::{Secret, SecretHash},
 };
 
-pub trait Ledger: swap_protocols::Ledger + ExtractSecret {
+pub trait Ledger: swap_protocols::Ledger {
     type LockDuration: PartialEq
         + Eq
         + Hash
@@ -23,8 +23,36 @@ pub trait Ledger: swap_protocols::Ledger + ExtractSecret {
         + PartialEq
         + Debug
         + Into<<Self as swap_protocols::ledger::Ledger>::Identity>;
+
+    fn extract_secret(
+        transaction: &RedeemTransaction<Self>,
+        secret_hash: &SecretHash,
+    ) -> Option<Secret>;
 }
 
-pub trait ExtractSecret: swap_protocols::ledger::Ledger {
-    fn extract_secret(txn: &Self::Transaction, secret_hash: &SecretHash) -> Option<Secret>;
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct FundTransaction<L: Ledger>(pub L::Transaction);
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct RedeemTransaction<L: Ledger>(pub L::Transaction);
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct RefundTransaction<L: Ledger>(pub L::Transaction);
+
+impl<L: Ledger> AsRef<L::Transaction> for FundTransaction<L> {
+    fn as_ref(&self) -> &L::Transaction {
+        &self.0
+    }
+}
+
+impl<L: Ledger> AsRef<L::Transaction> for RedeemTransaction<L> {
+    fn as_ref(&self) -> &L::Transaction {
+        &self.0
+    }
+}
+
+impl<L: Ledger> AsRef<L::Transaction> for RefundTransaction<L> {
+    fn as_ref(&self) -> &L::Transaction {
+        &self.0
+    }
 }
